+++ src/App.vue
@@ -5,6 +5,7 @@
     </button>
     <div id="nav">
       <router-link to="/">Home</router-link>|
+      <router-link to="/theme-toggling">Theme toggling</router-link>|
       <router-link to="/about">About</router-link>
     </div>
     <router-view />