+++ src/router/index.ts
@@ -1,6 +1,7 @@
 import Vue from "vue";
 import Router from "vue-router";
 import PdfViewer from "@/views/pdf-app.vue";
+import ThemeToggling from "@/views/ThemeToggling.vue";
 import About from "@/components/about.vue";
 
 Vue.use(Router);
@@ -13,6 +14,11 @@ export default new Router({
       path: "/",
       name: "home",
       component: PdfViewer
+    },
+    {
+      path: "/theme-toggling",
+      name: "theme-togglingh",
+      component: ThemeToggling
     },
     {
       path: "/about",